% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robocov_box.R
\name{Robocov_box}
\alias{Robocov_box}
\title{Robust correlation estimation using box constraints on Fisher Z-scores}
\usage{
Robocov_box(data_with_missing)
}
\arguments{
\item{data_with_missing}{Samples by features data matrix. May contain missing entries (NA) values.}
}
\description{
A robust estimation of correlation matrix for data with missing entries using box
constraint on the Fisher z-score values.
}
\examples{
data("sample_by_feature_data")
out = Robocov_box(sample_by_feature_data)
corrplot::corrplot(as.matrix(out), diag = FALSE,
        col = colorRampPalette(c("blue", "white", "red"))(200),
        tl.pos = "td", tl.cex = 0.4, tl.col = "black",
        rect.col = "white",na.label.col = "white",
        method = "color", type = "upper")

}
\keyword{box}
\keyword{correlation}
\keyword{shrinkage,}
