% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robocov_box_slack.R
\name{Robocov_box_slack}
\alias{Robocov_box_slack}
\title{Robust correlation estimation using box constraints on Fisher Z-scores  with slack variables}
\usage{
Robocov_box_slack(data_with_missing, alpha = 1)
}
\arguments{
\item{data_with_missing}{Samples by features data matrix. May contain missing entries (NA) values.}

\item{alpha}{The tuning parameter for the L-1 shrinkage of the slack variables.}
}
\description{
A robust estimation of correlation matrix for data with missing entries using box
constraint, alongside slack variables, on the Fisher z-score values with added L-1 penalty on the
slack variables.
}
\examples{
data("sample_by_feature_data")
out = Robocov_box_slack(sample_by_feature_data, alpha = 1)
corrplot::corrplot(as.matrix(out), diag = FALSE,
        col = colorRampPalette(c("blue", "white", "red"))(200),
        tl.pos = "td", tl.cex = 0.4, tl.col = "black",
        rect.col = "white",na.label.col = "white",
        method = "color", type = "upper")

}
\keyword{box}
\keyword{correlation}
\keyword{shrinkage,}
