% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robocov_cov.R
\name{Robocov_cov}
\alias{Robocov_cov}
\title{Robocov covariance estimation using box constraints on Fisher Z-scores}
\usage{
Robocov_cov(data_with_missing, loss = c("lasso", "ridge", "elasticnet"))
}
\arguments{
\item{data_with_missing}{Samples by features data matrix. May contain missing entries
(NA) values.}

\item{loss}{Specify if we minimize L-1 (`lasso`), L-2 (`ridge`) or elastic-net (`elasticnet`)
loss functions.}
}
\description{
A robust estimation of covariance matrix for data with missing entries using box
constraint on the difference between the population covariance matrix and pairwise sample covariance
matrix.
}
\examples{
data("sample_by_feature_data")
out = Robocov_cov(sample_by_feature_data)
corrplot::corrplot(as.matrix(cov2cor(out)), diag = FALSE,
        col = colorRampPalette(c("blue", "white", "red"))(200),
        tl.pos = "td", tl.cex = 0.4, tl.col = "black",
        rect.col = "white",na.label.col = "white",
        method = "color", type = "upper")

}
\keyword{box}
\keyword{correlation}
\keyword{shrinkage,}
