% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robocov_cov_slack.R
\name{Robocov_cov_slack}
\alias{Robocov_cov_slack}
\title{Robocov covariance estimation using box constraints on Fisher Z-scores with slack variables}
\usage{
Robocov_cov_slack(data_with_missing, alpha = 1, loss = c("lasso",
  "ridge", "elasticnet"))
}
\arguments{
\item{data_with_missing}{Samples by features data matrix. May contain missing entries (NA) values.}

\item{alpha}{The tuning parameter for the L-1 shrinkage of the slack variables.}

\item{loss}{Specify if we minimize L-1 (`lasso`), L-2 (`ridge`) or elastic-net (`elasticnet`)
loss functions.}
}
\description{
A robust estimation of covariance matrix for data with missing entries using box
constraint on the difference between the population covariance matrix and pairwise sample covariance
matrix, with added L-1 penalty on the slack variables. This is a more flexible version of the
Robocov_cov function.
}
\examples{
data("sample_by_feature_data")
out = Robocov_cov_slack(sample_by_feature_data, alpha = 1)
corrplot::corrplot(as.matrix(cov2cor(out)), diag = FALSE,
        col = colorRampPalette(c("blue", "white", "red"))(200),
        tl.pos = "td", tl.cex = 0.4, tl.col = "black",
        rect.col = "white",na.label.col = "white",
        method = "color", type = "upper")

}
\keyword{box}
\keyword{correlation}
\keyword{shrinkage,}
