% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robocov_local.R
\name{Robocov_local}
\alias{Robocov_local}
\title{Robust correlation estimation using local approximation to Fisher Z-scores}
\usage{
Robocov_local(data_with_missing, alpha = 1, loss = c("lasso", "ridge",
  "elasticnet"))
}
\arguments{
\item{data_with_missing}{Samples by features data matrix. May contain missing entries (NA) values.}

\item{alpha}{The tuning parameter for the L-1 scaling of the correlation values.}

\item{loss}{Specify if we minimize L-1 (`lasso`), L-2 (`ridge`) or elastic-net (`elasticnet`)
loss functions.}
}
\description{
A robust estimation of correlation matrix for data with missing entries using
a local second order Taylor series approximation of the Fisher z-score values.
}
\examples{
data("sample_by_feature_data")
out = Robocov_local(sample_by_feature_data, alpha = 1)
corrplot::corrplot(as.matrix(out), diag = FALSE,
        col = colorRampPalette(c("blue", "white", "red"))(200),
        tl.pos = "td", tl.cex = 0.4, tl.col = "black",
        rect.col = "white",na.label.col = "white",
        method = "color", type = "upper")

}
\keyword{box}
\keyword{correlation}
\keyword{shrinkage,}
