% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robocov_precision.R
\name{Robocov_precision}
\alias{Robocov_precision}
\title{Robust partial correlation estimation using Box constraint on Fisher Z-scores}
\usage{
Robocov_precision(data_with_missing, alpha, lambda = 1)
}
\arguments{
\item{data_with_missing}{The samples by features data matrix. May contain NA values.}

\item{alpha}{The tuning parameter for L-1 penalty.}

\item{lambda}{The weight of the constraint in the shrinkage. Default lambda taken to be 1.}
}
\description{
A robust estimation of partial correlation matrix for data with missing entries using box
constraint on Fisher Z-scores.
}
\examples{
data("sample_by_feature_data")
out = Robocov_precision(sample_by_feature_data, alpha = 0.1)
corrplot::corrplot(as.matrix(out), diag = FALSE,
        col = colorRampPalette(c("blue", "white", "red"))(200),
        tl.pos = "td", tl.cex = 0.4, tl.col = "black",
        rect.col = "white",na.label.col = "white",
        method = "color", type = "upper")

}
\keyword{box}
\keyword{correlation}
\keyword{shrinkage,}
